Extract helper delegate for maps using current location
